+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_wrong_names.py
@@ -1,9 +1,39 @@
 import pytest
 
 from wemake_python_styleguide.violations.naming import (
+    UpperCaseAttributeViolation,
     WrongVariableNameViolation,
 )
 from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+
+
+@pytest.mark.parametrize('wrong_name', [
+    'SOME',
+    'Value',
+    'nO',
+    'dATa',
+])
+def test_wrong_variable_name_case_insensitive(
+    assert_errors,
+    assert_error_text,
+    parse_ast_tree,
+    naming_template,
+    default_options,
+    mode,
+    wrong_name,
+):
+    """Ensures that wrong names are not allowed, case insensitive."""
+    tree = parse_ast_tree(mode(naming_template.format(wrong_name)))
+
+    visitor = WrongNameVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(
+        visitor,
+        [WrongVariableNameViolation],
+        ignored_types=(UpperCaseAttributeViolation,),
+    )
+    assert_error_text(visitor, wrong_name, multiple=True)
 
 
 @pytest.mark.parametrize('wrong_name', [