+++ CHANGELOG.md
@@ -99,6 +99,7 @@ character combination which is not easy to read
 - Fixes `WPS114` not to be so strict
 - Fixes `WPS122` not raising for `for` and `async for` definitions
 - Fixes `WPS400` raising for `# type: ignore[override]` comments
+- Fixes `WPS115` not raising for attributes inside other nodes
 
 ### Misc