+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Semantic versioning in our case means:
 - Minor releases do bring new features and configuration options. New violations can be added. Code that passes `x.0.y` might not pass on `x.1.y` release.
 - Major releases inidicate significant milestones or serious breaking changes.
 
+## 0.15.0 aka New runtime
+
+### Bugfixes
+
+- Fixes how wrong variable names were checked case sensitive with `WPS110`
 
 ## 0.14.0 aka The Walrus fighter